# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateMarket:
    properties:
      address:
        type: string
      name:
        type: string
      phone_number:
        type: string
      product_ids:
        items:
          type: string
        type: array
    type: object
...
  models.PatchRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
...
  models.UpdateMarket:
    properties:
      address:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      product_ids:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /v1/market:
    get:
      consumes:
      - application/json
      description: Get List Market
      operationId: get_list_market
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get List Market
      tags:
      - Market
    post:
      consumes:
      - application/json
      description: Create Market
      operationId: create_market
      parameters:
      - description: CreateMarketRequest
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/models.CreateMarket'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Market
      tags:
      - Market
  /v1/market/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Market
      operationId: delete_market
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Market
      tags:
      - Market
    get:
      consumes:
      - application/json
      description: Get By ID Market
      operationId: get_by_id_market
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Market
      tags:
      - Market
    patch:
      consumes:
      - application/json
      description: Patch Market
      operationId: patch_market
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: PatchMarketRequest
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/models.PatchRequest'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Patch Market
      tags:
      - Market
    put:
      consumes:
      - application/json
      description: Update Market
      operationId: update_market
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateMarketRequest
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMarket'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update Market
      tags:
      - Market
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"